# Speed up bulk upload with missing memoization