--- a/app/services/bulk_upload/lettings/year2022/row_parser.rb
+++ b/app/services/bulk_upload/lettings/year2022/row_parser.rb
@@ -340,9 +340,9 @@ class BulkUpload::Lettings::Year2022::RowParser
 
     return true if blank_row?
 
-    super
-
     log.valid?
+
+    super
 
     log.errors.each do |error|
       fields = field_mapping_for_errors[error.attribute] || []
@@ -650,7 +650,7 @@ private
   end
 
   def questions
-    log.form.subsections.flat_map { |ss| ss.applicable_questions(log) }
+    @questions ||= log.form.subsections.flat_map { |ss| ss.applicable_questions(log) }
   end
 
   def validate_nulls