# Context

- Bulk upload has been acting slow and spent some time actually investigating why

# Changes

- Add missing memoization to bulk upload which causes performance boost
- See evidence below

```
Finished in 4 minutes 0.9 seconds (files took 5.27 seconds to load)
210 examples, 0 failures, 2 pending

Finished in 30.85 seconds (files took 5.23 seconds to load)
210 examples, 0 failures, 2 pending
```