@arfacamble 2023-04-14
I wondered when I saw the lines that failed the linter if that was best practice, but I also know that chaining on a do end is also gross.
Interested to know what is the nicest way to write that.